+++ Core/BlazorMvvm.Core/Attributes/BlazorObservablePropertyAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace BlazorMvvm
+{
+    [AttributeUsage(AttributeTargets.Field)]
+    public class BlazorObservablePropertyAttribute : Attribute
+    {
+    }
+}